Rename tokenizers to token extractors
